test(json-schema): split async scanner tests to sibling spec (resolve FIX-003)
Move the JsonSchemaScannerAsync describe block (6 cases) from JsonSchemaScanner.test.ts to JsonSchemaScannerAsync.spec.ts, bringing both files under the FCA 3+12 rule (test.ts 17 -> 11, spec.ts 7 -> 13). Pure move, no test logic changed; 378/378 package tests still pass. Removes the now-resolved FIX-003 debt record (was deferred during resolve --auto as pre-existing/out-of-scope; addressed on request).
